changed all physics values to double

+++ src/texts.hpp
@@ -7,13 +7,13 @@
 
 class Texts : public Drawable
 {
-	std::vector<std::pair<glm::vec2, std::string>> to_draw2d;
-	std::vector<std::pair<glm::vec3, std::string>> to_draw3d;
+	std::vector<std::pair<glm::dvec2, std::string>> to_draw2d;
+	std::vector<std::pair<glm::dvec3, std::string>> to_draw3d;
 
 public:
 
-	void addText(float x, float y, const std::string& text);
-	void addText(glm::vec3, const std::string& text);
+	void addText(double x, double y, const std::string& text);
+	void addText(glm::dvec3, const std::string& text);
 
 	void clear();
